Делает дз модуляция
